+++ python/tests/test_reader.py
@@ -105,29 +105,13 @@ def test_reader_deser(dummy_reader_plugin):
 
 
 def test_reader_timeout(dummy_reader_plugin):
-    with sinetstream.MessageReader(service, topic, receive_timeout_ms=1000) as f:
+    with sinetstream.MessageReader(service, topic, receive_timeout_ms=3000) as f:
         pass
-    assert True
 
 
 def test_reader_kafka_opt(dummy_reader_plugin):
     with sinetstream.MessageReader(service, topic, heartbeat_interval_ms=1000) as f:
         pass
-    assert True
-
-
-@pytest.mark.skip
-def test_reader_timeout():
-    tout = 100
-    t1 = time.time()
-    with sinetstream.MessageReader(service, topic,  receive_timeout_ms=tout) as f:
-        for msg in f:
-            assert False
-    t2 = time.time()
-    t = (t2 - t1) * 1000.0
-    jitter_factor = 5
-    assert t > tout / jitter_factor
-    assert t < tout * jitter_factor
     assert True
 
 
@@ -143,3 +127,61 @@ def test_reader_seek():
         except AssertionError:
             pass  # If any partition is not currently assigned, or if no partitions are assigned.
     assert True
+
+
+def test_reader_topics_in_config_file(dummy_reader_plugin):
+    with sinetstream.MessageReader(service + "-topics") as f:
+        assert f.topics == topic
+
+
+def test_reader_topics_list_in_config_file(dummy_reader_plugin):
+    with sinetstream.MessageReader(service + "-topics-list") as f:
+        assert f.topics == [topic, "mss-test-002"]
+
+
+def test_reader_topic_in_config_file(dummy_reader_plugin):
+    with sinetstream.MessageReader(service + "-topic") as f:
+        assert f.topics == topic
+
+
+def test_reader_topic_list_in_config_file(dummy_reader_plugin):
+    with sinetstream.MessageReader(service + "-topic-list") as f:
+        assert f.topics == [topic, "mss-test-002"]
+
+
+def test_reader_topics_in_config_file_and_arg(dummy_reader_plugin):
+    topic2 = "mss-test-002"
+    with sinetstream.MessageReader(service + "-topic", topic2) as f:
+        assert f.topics == topic2
+
+
+def test_reader_topics_in_config_file_and_kwarg(dummy_reader_plugin):
+    service_name = service + "-topic"
+    topic2 = "mss-test-002"
+    with sinetstream.MessageReader(topics=topic2, service=service_name) as f:
+        assert f.topics == topic2
+
+
+def test_reader_no_topics(dummy_reader_plugin):
+    with pytest.raises(sinetstream.InvalidArgumentError):
+        with sinetstream.MessageReader(service) as f:
+            pass
+
+
+def test_reader_empty_topics_list(dummy_reader_plugin):
+    with pytest.raises(sinetstream.InvalidArgumentError):
+        with sinetstream.MessageReader(service, topics=[]) as f:
+            pass
+
+
+def test_open_close(dummy_reader_plugin):
+    f = sinetstream.MessageReader(service, topic).open()
+    assert hasattr(f, '__iter__')
+    f.close()
+
+
+def test_close_twice(dummy_reader_plugin):
+    f = sinetstream.MessageReader(service, topic).open()
+    f.close()
+    with pytest.raises(sinetstream.SinetError):
+        f.close()